requirements.txt
pandas
numpy
requests
pyyaml
streamlit
python-dotenv
openai
